Simplify logic and improve warnings regarding number of labels/legends/titles
When automatic subplots are 'off', then only the first title and label
are used. Thus, we print a warning if users provide more than one title
or label.

When creating a phase plot, only the first legend label is used, so we
print a warning if multiple legend labels are given.

For each warning, we add tests to verify that the warnings are shown at
the approriate time.

The logic aroudn selecting which label, legend label, or title was more
complicated than necessary, so this commit simplifies it.