New functions for faster execution
I've created a new variant of HyEQsolver based on the old v2 code that runs faster than the version on v3 . Instead of allocating memory on every iteration, memory is allocated in advance. Speed up is significant in cases when there is a large amount of jumps (like sampled systems) and I've included a simple example to showcase its effectiveness.
I've also tried to implement the same trick with the new code, but for some reason it is not working. I've managed to pinpoint the slowdown to line 408. Identical variable assignment works perfectly with the old code version, so I am not sure what is the problem. Maybe together we can brainstorm a solution.